ui: refactor schema insights to remove redux and sagas
Refactors the schema insights component to no longer
require redux and sagags. It now uses `useSwr` for data
fetching. In order to remove the dependency on redux
and sagas, some functionality was removed from this
component. This functionality is:
  1. dispatching analytics.track actions
  2. persisting table sorting and filtering to browser
     session storage.

The analytics.track actions were used in cloud console
to track when the table is sorted or a filter is applied.
We decided this doesn't provide very much value to
replace and removed it.

Epic: None
Release note: None